refactor: RealEstate 엔티티 수정
1. 모든 필드가 wrapper class를 사용하도록 수정
2. pyeongsu 필드의 자료형을 Integer로 변경
3. Builder 애노테이션 대상을 id값이 포함되지 않은 생성자로 변경

+++ src/main/java/com/project/zipkok/dto/GetRealEstateResponse.java
@@ -41,7 +41,7 @@ public class GetRealEstateResponse {
 
     private Float areaSize;
 
-    private Long pyeongsu;
+    private Integer pyeongsu;
 
     private String realEstateType;
